+++ Lost Together/Assets/InputSystemActions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e2e2b8be03cf4cefbcee09cf4ad6ec3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
