+++ Lost Together/Assets/InputSystemActions/Player.inputactions.meta
@@ -1,0 +1,14 @@
+fileFormatVersion: 2
+guid: 18dc0b58a0a82489bb44c62f3b2f2eb4
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
+  generateWrapperCode: 0
+  wrapperCodePath: 
+  wrapperClassName: 
+  wrapperCodeNamespace:
