+++ Lost Together/Assets/InputSystemActions/Player2.inputactions.meta
@@ -1,0 +1,14 @@
+fileFormatVersion: 2
+guid: 6cb4d9268c7d94398ba7edc78da36f95
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
+  generateWrapperCode: 0
+  wrapperCodePath: 
+  wrapperClassName: 
+  wrapperCodeNamespace:
