+++ Lost Together/Assets/Scripts/Player1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da8e167239c9044fc8ff71a4cbdc439a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
