+++ Lost Together/Assets/Scripts/Player2.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d67deacbc8c254146b619cae653de7a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
