+++ Lost Together/Assets/Scripts/Player2/Movement_P2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f5ff7d50fa704dd7a2fb7e86fce5744
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
